# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "马森雨个人技术博客（masenyu.top）RESTful API 接口文档，涵盖文章、分类、标签、项目、友链、评论、AI助手、数据看板与站点管理等模块。",
        "title": "马森雨个人技术博客 API",
        "contact": {
            "name": "马森雨",
            "url": "https://masenyu.top"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/admin/2fa/disable": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CsrfToken": []
                    }
                ],
                "description": "校验当前密码及动态码，关闭管理员 2FA 功能",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员认证与安全"
                ],
                "summary": "关闭两步验证",
                "parameters": [
                    {
                        "description": "密码与动态码",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.Disable2FARequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "关闭成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "密码或验证码错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/2fa/enable": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CsrfToken": []
                    }
                ],
                "description": "输入 TOTP 验证码确认绑定，正式激活 2FA",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员认证与安全"
                ],
                "summary": "校验并启用两步验证",
                "parameters": [
                    {
                        "description": "TOTP 验证码",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.Enable2FARequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "启用成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "验证码错误或未生成密钥",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/2fa/setup": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CsrfToken": []
                    }
                ],
                "description": "为当前管理员生成未激活的 TOTP 密钥和二维码配置 URL",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员认证与安全"
                ],
                "summary": "生成 2FA 密钥与配置二维码",
                "responses": {
                    "200": {
                        "description": "生成成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.Setup2FAResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "两步验证已启用",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/admin/login": {
            "post": {
                "description": "校验管理员账号与密码，若启用 2FA 则返回 requires2FA 标志并下发临时待验证凭据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员认证与安全"
                ],
                "summary": "管理员登录",
                "parameters": [
                    {
                        "description": "管理员登录表单",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功或要求2FA",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.LoginResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "账号或密码错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "尝试过于频繁",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/login/2fa": {
            "post": {
                "description": "提交动态验证码完成 2FA 登录，验证成功后签发管理端 Token 与 CSRF Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员认证与安全"
                ],
                "summary": "2FA 二次验证登录",
                "parameters": [
                    {
                        "description": "TOTP 6位验证码",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.Login2FARequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "2FA 验证成功并完成登录",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.LoginResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "验证码错误或临时凭据失效",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CsrfToken": []
                    }
                ],
                "description": "清除管理员 Session Cookie 与 CSRF 令牌",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员认证与安全"
                ],
                "summary": "管理员退出登录",
                "responses": {
                    "200": {
                        "description": "退出成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/admin/password": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CsrfToken": []
                    }
                ],
                "description": "校验原密码并更新管理员账号密码，更新成功后使旧会话失效",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员认证与安全"
                ],
                "summary": "管理员修改密码",
                "parameters": [
                    {
                        "description": "修改密码请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "修改密码成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误或密码不合规",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/admin/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CsrfToken": []
                    }
                ],
                "description": "获取当前已登录管理员的个人信息、权限及2FA状态",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员认证与安全"
                ],
                "summary": "获取当前管理员资料",
                "responses": {
                    "200": {
                        "description": "个人资料",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.AdminUserDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.AdminUserDTO": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "totpEnabled": {
                    "type": "boolean"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "handler.ChangePasswordRequest": {
            "type": "object",
            "properties": {
                "currentPassword": {
                    "type": "string"
                },
                "newPassword": {
                    "type": "string"
                }
            }
        },
...
        "handler.Disable2FARequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "handler.Enable2FARequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                }
            }
        },
...
        "handler.Login2FARequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                }
            }
        },
        "handler.LoginRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "handler.LoginResponse": {
            "type": "object",
            "properties": {
                "requires2FA": {
                    "type": "boolean"
                },
                "user": {
                    "$ref": "#/definitions/handler.AdminUserDTO"
                }
            }
        },
...
        "handler.Setup2FAResponse": {
            "type": "object",
            "properties": {
                "otpauthUrl": {
                    "type": "string"
                },
                "secret": {
                    "type": "string"
                }
            }
        },
...
        "response.Envelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "参数错误"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "请输入 JWT Token，格式为: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CsrfToken": {
            "description": "管理端防跨站请求伪造令牌",
            "type": "apiKey",
            "name": "X-CSRF-Token",
            "in": "header"
        }
    }
}